Fix: Temporarily disable vince converter for .vcf previews
The vince converter, used for previewing .vcf vCard files via the VinceLib library,
was calling the `unloadable` method. This method was deprecated in Rails 5 and
removed in Rails 6. As Redmine 6.0.5 is based on Rails 7.2.2.1, the `unloadable`
method is no longer available, leading to a NameError.
This commit temporarily disables the vince converter to allow plugin migration
and asset precompilation to proceed.